changed 'int' to 'unit32_t' or similar

+++ src/mcu/include/GenerateFrame.h
@@ -55,7 +55,7 @@ public:
      * @param frameType The type of the frame (default: DATA_FRAME).
      * @return Returns 0 on success, -1 on failure.
      */
-    int SendFrame(int can_id, std::vector<int> data, FrameType frameType = DATA_FRAME);
+    int SendFrame(uint32_t can_id, std::vector<uint8_t> data, FrameType frameType = DATA_FRAME);
 
     /**
      * @brief Gets the CAN frame.
@@ -76,7 +76,7 @@ public:
      * @param frameType The type of the frame (default: DATA_FRAME).
      * @return Returns the created CAN frame.
      */
-    can_frame CreateFrame(int can_id, std::vector<int> data, FrameType frameType = DATA_FRAME);
+    can_frame CreateFrame(uint32_t can_id, std::vector<uint8_t> data, FrameType frameType = DATA_FRAME);
 
     /**
      * @brief Creates a CAN frame with multiple segments with the given arguments.
@@ -86,7 +86,7 @@ public:
      * @param response The response data.
      * @param firstFrame Indicates if it is the first frame (default: true).
      */
-    void CreateFrameLong(int can_id, int sid, int dataIdentifier, std::vector<int> response, bool firstFrame);
+    void CreateFrameLong(uint32_t can_id, uint8_t sid, uint8_t dataIdentifier, std::vector<uint8_t> response, bool firstFrame);
 
     /**
      * @brief Performs session control.
@@ -94,14 +94,14 @@ public:
      * @param subfunction The subfunction.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void SessionControl(int can_id, int subfunction, bool response = false);
+    void SessionControl(uint32_t can_id, uint8_t subfunction, bool response = false);
 
     /**
      * @brief Performs ECU reset.
      * @param can_id The CAN ID of the frame.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void EcuReset(int can_id, bool response = false);
+    void EcuReset(uint32_t can_id, bool response = false);
 
     /**
      * @brief Reads data by identifier.
@@ -109,7 +109,7 @@ public:
      * @param dataIdentifier The data identifier.
      * @param response The response data (default: empty vector).
      */
-    void ReadDataByIdentifier(int can_id, int dataIdentifier, std::vector<int> response = {});
+    void ReadDataByIdentifier(uint32_t can_id, uint8_t dataIdentifier, std::vector<uint8_t> response = {});
 
     /**
      * @brief Reads data by identifier (long format).
@@ -118,27 +118,27 @@ public:
      * @param response The response data (default: empty vector).
      * @param firstFrame Indicates if it is the first frame (default: true).
      */
-    void ReadDataByIdentifierLong(int can_id, int dataIdentifier, std::vector<int> response = {}, bool firstFrame = true);
+    void ReadDataByIdentifierLong(uint32_t can_id, uint8_t dataIdentifier, std::vector<uint8_t> response = {}, bool firstFrame = true);
 
     /**
      * @brief Sends flow control frame.
      * @param can_id The CAN ID of the frame.
      */
-    void FlowControlFrame(int can_id);
+    void FlowControlFrame(uint32_t can_id);
 
     /**
      * @brief Sends authentication request seed frame.
      * @param can_id The CAN ID of the frame.
      * @param seed The seed data (default: empty vector).
      */
-    void AuthenticationRequestSeed(int can_id, const std::vector<int> &seed = {});
+    void AuthenticationRequestSeed(uint32_t can_id, const std::vector<uint8_t> &seed = {});
 
     /**
      * @brief Sends authentication send key frame.
      * @param can_id The CAN ID of the frame.
      * @param key The key data (default: empty vector).
      */
-    void AuthenticationSendKey(int can_id, const std::vector<int> &key = {});
+    void AuthenticationSendKey(uint32_t can_id, const std::vector<uint8_t> &key = {});
 
     /**
      * @brief Performs routine control.
@@ -147,14 +147,14 @@ public:
      * @param routine_identifier The routine identifier.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void RoutineControl(int can_id, int subfunction, int routine_identifier, bool response = false);
+    void RoutineControl(uint32_t can_id, uint8_t subfunction, uint8_t routine_identifier, bool response = false);
 
     /**
      * @brief Sends tester present frame.
      * @param can_id The CAN ID of the frame.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void TesterPresent(int can_id, bool response = false);
+    void TesterPresent(uint32_t can_id, bool response = false);
 
     /**
      * @brief Reads memory by address.
@@ -163,7 +163,7 @@ public:
      * @param memory_address The memory address.
      * @param response The response data (default: empty vector).
      */
-    void ReadMemoryByAddress(int can_id, int memory_size, int memory_address, std::vector<int> response = {});
+    void ReadMemoryByAddress(uint32_t can_id, uint8_t memory_size, uint8_t memory_address, std::vector<uint8_t> response = {});
 
     /**
      * @brief Reads memory by address (long format).
@@ -173,7 +173,7 @@ public:
      * @param response The response data.
      * @param first_frame Indicates if it is the first frame (default: true).
      */
-    void ReadMemoryByAddressLong(int can_id, int memory_size, int memory_address, std::vector<int> response, bool first_frame = true);
+    void ReadMemoryByAddressLong(uint32_t can_id, uint8_t memory_size, uint8_t memory_address, std::vector<uint8_t> response, bool first_frame = true);
 
     /**
      * @brief Writes data by identifier.
@@ -181,7 +181,7 @@ public:
      * @param identifier The identifier.
      * @param data_parameter The data parameter (default: empty vector).
      */
-    void WriteDataByIdentifier(int can_id, int identifier, std::vector<int> data_parameter = {});
+    void WriteDataByIdentifier(uint32_t can_id, uint8_t identifier, std::vector<uint8_t> data_parameter = {});
 
     /**
      * @brief Writes data by identifier (long format).
@@ -190,7 +190,7 @@ public:
      * @param data_parameter The data parameter.
      * @param first_frame Indicates if it is the first frame (default: true).
      */
-    void WriteDataByIdentifierLong(int can_id, int identifier, std::vector<int> data_parameter, bool first_frame = true);
+    void WriteDataByIdentifierLong(uint32_t can_id, uint8_t identifier, std::vector<uint8_t> data_parameter, bool first_frame = true);
 
     /**
      * @brief Reads DTC information.
@@ -198,7 +198,7 @@ public:
      * @param subfunction The subfunction.
      * @param dtc_status_mask The DTC status mask.
      */
-    void ReadDtcInformation(int can_id, int subfunction, int dtc_status_mask);
+    void ReadDtcInformation(uint32_t can_id, uint8_t subfunction, uint8_t dtc_status_mask);
 
     /**
      * @brief Sends DTC information response frame (format 01).
@@ -207,7 +207,7 @@ public:
      * @param dtc_format_identifier The DTC format identifier.
      * @param dtc_count The DTC count.
      */
-    void ReadDtcInformationResponse01(int can_id, int status_availability_mask, int dtc_format_identifier, int dtc_count);
+    void ReadDtcInformationResponse01(uint32_t can_id, uint8_t status_availability_mask, uint8_t dtc_format_identifier, uint8_t dtc_count);
 
     /**
      * @brief Clears diagnostic information.
@@ -215,7 +215,7 @@ public:
      * @param group_of_dtc The group of DTC (default: {0xFF, 0xFF, 0xFF}).
      * @param response Indicates if it is a response frame (default: false).
      */
-    void ClearDiagnosticInformation(int can_id, std::vector<int> group_of_dtc = {0xFF, 0xFF, 0xFF}, bool response = false);
+    void ClearDiagnosticInformation(uint32_t can_id, std::vector<uint8_t> group_of_dtc = {0xFF, 0xFF, 0xFF}, bool response = false);
 
     /**
      * @brief Accesses timing parameters.
@@ -223,7 +223,7 @@ public:
      * @param subfunction The subfunction.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void AccessTimingParameters(int can_id, int subfunction, bool response = false);
+    void AccessTimingParameters(uint32_t can_id, uint8_t subfunction, bool response = false);
 
     /**
      * @brief Sends negative response frame.
@@ -231,7 +231,7 @@ public:
      * @param sid The service identifier.
      * @param nrc The negative response code.
      */
-    void NegativeResponse(int can_id, int sid, int nrc);
+    void NegativeResponse(uint32_t can_id, uint8_t sid, uint8_t nrc);
 
     /**
      * @brief Requests download.
@@ -240,14 +240,14 @@ public:
      * @param memory_address The memory address.
      * @param memory_size The memory size.
      */
-    void RequestDownload(int can_id, int data_format_identifier, int memory_address, int memory_size);
+    void RequestDownload(uint32_t can_id, uint8_t data_format_identifier, uint8_t memory_address, uint8_t memory_size);
 
     /**
      * @brief Sends request download response frame.
      * @param can_id The CAN ID of the frame.
      * @param max_number_block The maximum number of blocks.
      */
-    void RequestDownloadResponse(int can_id, int max_number_block);
+    void RequestDownloadResponse(uint32_t can_id, uint8_t max_number_block);
 
     /**
      * @brief Transfers data.
@@ -255,7 +255,7 @@ public:
      * @param block_sequence_counter The block sequence counter.
      * @param transfer_request The transfer request (default: empty vector).
      */
-    void TransferData(int can_id, int block_sequence_counter, std::vector<int> transfer_request = {});
+    void TransferData(uint32_t can_id, uint8_t block_sequence_counter, std::vector<uint8_t> transfer_request = {});
 
     /**
      * @brief Transfers data (long format).
@@ -264,14 +264,14 @@ public:
      * @param transfer_request The transfer request.
      * @param first_frame Indicates if it is the first frame (default: true).
      */
-    void TransferDataLong(int can_id, int block_sequence_counter, std::vector<int> transfer_request, bool first_frame = true);
+    void TransferDataLong(uint32_t can_id, uint8_t block_sequence_counter, std::vector<uint8_t> transfer_request, bool first_frame = true);
 
     /**
      * @brief Requests transfer exit.
      * @param can_id The CAN ID of the frame.
      * @param response Indicates if it is a response frame (default: false).
      */
-    void RequestTransferExit(int can_id, bool response = false);
+    void RequestTransferExit(uint32_t can_id, bool response = false);
 
 private:
     int socket = -1;
@@ -284,7 +284,7 @@ private:
      * @param index The index to insert the bytes.
      * @param num_bytes The number of bytes to insert.
      */
-    void InsertBytes(std::vector<int> &data, unsigned int index, int num_bytes);
+    void InsertBytes(std::vector<uint8_t> &data, uint8_t index, uint8_t num_bytes);
 
     /**
      * @brief Counts the number of digits in a number.
